fix: if the isalpha() and isdigit() functions are implemented with macros, it is possible that a warning is raised by gcc when a 'char' is passed to these functions instead of an 'int'. A 'char' will not be typecasted automatically to an 'int' when macros are used. This results in a warning "invalid type used in array subscript".

+++ src/libgds/params.h
@@ -80,7 +80,7 @@ typedef enum {
  */
 static inline int _is_param_first_char(char c)
 {
-  return (isalpha(c) || (c == PARAM_CHAR_UNDERSCORE));
+  return (isalpha(((int) c)) || (c == PARAM_CHAR_UNDERSCORE));
 }
 
 // -----[ _is_param_char ]-------------------------------------------
@@ -89,7 +89,7 @@ static inline int _is_param_first_char(char c)
  */
 static inline int _is_param_char(char c)
 {
-  return (_is_param_first_char(c) || isdigit(c));
+  return (_is_param_first_char(c) || isdigit(((int) c)));
 }
 
 #ifdef __cplusplus
